moved `load-http` into `resolve`
