+++ backend/internal/compilerServer/compiler.go
@@ -17,6 +17,7 @@ import (
 
 	es "github.com/evanw/esbuild/pkg/api"
 	"robinplatform.dev/internal/compile/buildError"
+	"robinplatform.dev/internal/compile/resolve"
 	"robinplatform.dev/internal/compile/toolkit"
 	"robinplatform.dev/internal/identity"
 	"robinplatform.dev/internal/log"
@@ -366,7 +367,7 @@ func (app *CompiledApp) buildClientJs() error {
 		Plugins: concat(
 			getExtractServerPlugins(appConfig, app),
 			toolkit.Plugin(appConfig),
-			[]es.Plugin{esbuildPluginLoadHttp},
+			resolve.HttpPlugin(httpClient),
 			getResolverPlugins(appConfig, pagePath),
 			getCssLoaderPlugins(appConfig),
 
@@ -457,7 +458,7 @@ func (app *CompiledApp) buildServerBundle() error {
 		Define:   app.getEnvConstants(),
 		Plugins: concat(
 			[]es.Plugin{esbuildPluginMarkBuiltinsAsExternal},
-			[]es.Plugin{esbuildPluginLoadHttp},
+			resolve.HttpPlugin(httpClient),
 			[]es.Plugin{
 				{
 					Name: "resolve-abs-paths",
